refactor: 포지션, 스택 선택 ㅇ컴포넌트 분리
